Remove raw_window_handle feature
The main point of being able to get the window handle is so you can do platform-specification actions such as setting the window parent when spawning a file dialog. I realize you can get the raw window handle without needing to implement it library-side. I'm going to keep the example because it shows how to do this.